move util functions into utils.go
